Initial version - setup a file-push architecture for generating embeddings with Khoj
